fix(Refatoracao): Reorganizado o código da integração, alterado o README e adicionados os scripts dos bancos de dados
