+++ src/Integration/Database/Entities/produtos.entity.js
@@ -1,5 +1,5 @@
 const Sequelize = require('sequelize');
-const database = require('../postgresDatabase.provider');
+const database = require('../postgreSQL.provider');
 const Estoque = require('./estoque.entity');
  
 const Produto = database.define('Produtos', {
